// src/lib/types.ts
import { EventStore, ResourceModel } from '@bryntum/schedulerpro-thin';
import { PanelConfig, PanelListeners } from '@bryntum/core-thin';
import { TimeAxis } from '@bryntum/scheduler-thin';

import Task from '../views/Planning/partials/Task';

// MapPanel types
export interface MarkerConfig {
  color: string
}

export interface PopupConfig {
  offset: [number, number]
}

// Bryntum types
type constructParamsListeners = PanelListeners & { markerclick: (params: { eventRecord: Task }) => void };

export type constructParams = PanelConfig & {
  appendTo: string
  eventStore: EventStore
  flex: number
  listeners: constructParamsListeners
  ref: string
  timeAxis: TimeAxis
}

// Extend Bryntum types
export type DragDropContext = {
  // Base DragHelper context properties
  element: HTMLElement
  target: HTMLElement
  grabbed: HTMLElement
  relatedElements: HTMLElement[]
  valid: boolean
  newX: number
  newY: number

  // Custom properties
  task: Task
  totalDuration: number
  resourceRecord?: ResourceModel
}